Two new features: all variable "container" objects that can hold Var's have grown a 'v' attibute; from this attribute, you can create direct refs to Vars using namespace model classes, instances of those classes, or context expressions, and the value will get picked up and used in other models (this lets you do things like use a ref to a namespace Var as a value in an InfraModel, such as an IP address). Also, the 'nexus' has been beefed up with linkages to models in a model set, and a ref to the nexus has been placed directly into the ContextExpr object, so finding other models is as easy as 'ctxt.nexus.inf' from the NamespaceModel.
NOTE: I'm seeing infrequent failures of test014 in the ansible_tests.py
module. This only happens once in a blue moon, and never twice in a row.
I'm not convinced that it doesn't have something to do with a timeout in
the underlying SSH infra after the system has been idle or has just been
rebooted. Should keep an eye on it.